Create README.md
The dataset is about Netflix movies and TV shows. I have cleaned the data and then performed an exploratory analysis of the data; through these steps, we can answer the questions proposed in the pdf. I have used different graphics to be able to visualize the results obtained in different ways.